+++ src/services/resources/CRUDInterface.php
@@ -45,15 +45,13 @@ interface CRUDInterface
      * @param Objects $field
      * @param ConnectionInterface|string|null $connection
      * @param CacheInterface|string|null $cache
-     * @param TransformerCollectionInterface|array|null $transformer
      * @return bool
      */
     public function syncDown(
         ElementInterface $element,
         Objects $field,
         ConnectionInterface $connection = null,
-        CacheInterface $cache = null,
-        TransformerCollectionInterface $transformer = null
+        CacheInterface $cache = null
     ): bool;
 
     /**
@@ -61,15 +59,13 @@ interface CRUDInterface
      * @param Objects $field
      * @param ConnectionInterface|string|null $connection
      * @param CacheInterface|string|null $cache
-     * @param TransformerCollectionInterface|array|null $transformer
      * @return false|string
      */
     public function syncUp(
         ElementInterface $element,
         Objects $field,
         ConnectionInterface $connection = null,
-        CacheInterface $cache = null,
-        TransformerCollectionInterface $transformer = null
+        CacheInterface $cache = null
     ): bool;
 
     /**